Unified ItemView and ItemViewSelector
A new constructor that takes an ItemViewArg simplifies databinding since
you don't need two seperate methods for ItemView and ItemViewSelector.
a couple of BindingConversions keeps this backwards-compatible.

Note: If you are using a custom BindingCollectionAdapter, you should add
a cosntructor to create it with an ItemViewArg (the common case would
just be to call super). There is currently a fallback so that this
change is backwards-compatible but it will be removed in a future
update.