Use Gradle plugin 3.1.0-alpha6
Use Gradle plugin 3.1.0-alpha6 to be compatible with the new changes (databinding v2 and LiveData support). You need to recompile with this gradle plugin version, otherwise people won't be able to use your library when they build with this version. It leads to bad serialVersionUIDs and missing databinding generated classes).