ItemBinding can be contravariant in adapters (#165)
ItemBinding is used to map items of type T to their layout and
binding variable, among others. It might well be the case that
the passed ItemBinding can not only map instances of type T, but
any instances of some supertype of T, as well as T itself. This
relaxation is in fact already made for the ItemIds callback.